Extract enrichment type validation
